# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Barber Recall CRM API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /appointments:
    get:
      operationId: listAppointments
      tags: [appointments]
      summary: List appointments
      parameters:
        - name: date
          in: query
          schema:
            type: string
            format: date
        - name: barberId
          in: query
          schema:
            type: integer
        - name: clientId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Appointment list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Appointment"
    post:
      operationId: createAppointment
      tags: [appointments]
      summary: Create an appointment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppointmentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
  /appointments/{id}:
    get:
      operationId: getAppointment
      tags: [appointments]
      summary: Get an appointment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Appointment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
    patch:
      operationId: updateAppointment
      tags: [appointments]
      summary: Update an appointment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppointmentUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
    delete:
      operationId: deleteAppointment
      tags: [appointments]
      summary: Delete an appointment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── Barbers ─────────────────────────────────────────────────────────────────
  /barbers:
    get:
      operationId: listBarbers
      tags: [barbers]
      summary: List all barbers
      responses:
        "200":
          description: Barber list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Barber"
    post:
      operationId: createBarber
      tags: [barbers]
      summary: Create a barber
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BarberInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Barber"
  /barbers/{id}:
    patch:
      operationId: updateBarber
      tags: [barbers]
      summary: Update a barber
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BarberUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Barber"
    delete:
      operationId: deleteBarber
      tags: [barbers]
      summary: Delete a barber
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── Services ────────────────────────────────────────────────────────────────
...
  /campaigns:
    get:
      operationId: listCampaigns
      tags: [campaigns]
      summary: List all campaigns
      responses:
        "200":
          description: Campaign list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Campaign"
    post:
      operationId: createCampaign
      tags: [campaigns]
      summary: Create a campaign
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CampaignInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
  /campaigns/{id}:
    patch:
      operationId: updateCampaign
      tags: [campaigns]
      summary: Update a campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CampaignUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
    delete:
      operationId: deleteCampaign
      tags: [campaigns]
      summary: Delete a campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /campaigns/{id}/toggle:
    post:
      operationId: toggleCampaign
      tags: [campaigns]
      summary: Toggle campaign active state
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"

  # ─── Coupons ─────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Barber:
      type: object
      required: [id, nome, ativo, createdAt]
      properties:
        id:
          type: integer
        nome:
          type: string
        telefone:
          type: ["string", "null"]
        ativo:
          type: boolean
        foto:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    BarberInput:
      type: object
      required: [nome]
      properties:
        nome:
          type: string
        telefone:
          type: string
        foto:
          type: string
    BarberUpdate:
      type: object
      properties:
        nome:
          type: string
        telefone:
          type: string
        ativo:
          type: boolean
        foto:
          type: string

    # ─── Service ────────────────────────────────────────────────────
...
    Appointment:
      type: object
      required: [id, clienteId, valor, valorFinal, data, createdAt]
      properties:
        id:
          type: integer
        clienteId:
          type: integer
        clienteNome:
          type: ["string", "null"]
        barbeiroId:
          type: ["integer", "null"]
        barbeiroNome:
          type: ["string", "null"]
        servicoId:
          type: ["integer", "null"]
        servicoNome:
          type: ["string", "null"]
        valor:
          type: number
        desconto:
          type: number
        valorFinal:
          type: number
        data:
          type: string
          format: date-time
        observacoes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    AppointmentInput:
      type: object
      required: [clienteId, valor, valorFinal, data]
      properties:
        clienteId:
          type: integer
        barbeiroId:
          type: integer
        servicoId:
          type: integer
        valor:
          type: number
        desconto:
          type: number
        valorFinal:
          type: number
        data:
          type: string
          format: date-time
        observacoes:
          type: string
        couponCode:
          type: string
    AppointmentUpdate:
      type: object
      properties:
        barbeiroId:
          type: integer
        servicoId:
          type: integer
        valor:
          type: number
        desconto:
          type: number
        valorFinal:
          type: number
        data:
          type: string
          format: date-time
        observacoes:
          type: string

    # ─── Campaign ───────────────────────────────────────────────────
    Campaign:
      type: object
      required: [id, nome, tipo, dias, mensagem, ativo, createdAt]
      properties:
        id:
          type: integer
        nome:
          type: string
        tipo:
          type: string
          enum: [return, birthday, loyalty, custom]
        dias:
          type: integer
        mensagem:
          type: string
        cupomId:
          type: ["integer", "null"]
        cupomCodigo:
          type: ["string", "null"]
        ativo:
          type: boolean
        notificacoesEnviadas:
          type: integer
        createdAt:
          type: string
          format: date-time
    CampaignInput:
      type: object
      required: [nome, tipo, dias, mensagem]
      properties:
        nome:
          type: string
        tipo:
          type: string
          enum: [return, birthday, loyalty, custom]
        dias:
          type: integer
        mensagem:
          type: string
        cupomId:
          type: integer
    CampaignUpdate:
      type: object
      properties:
        nome:
          type: string
        tipo:
          type: string
          enum: [return, birthday, loyalty, custom]
        dias:
          type: integer
        mensagem:
          type: string
        cupomId:
          type: integer
        ativo:
          type: boolean

    # ─── Coupon ─────────────────────────────────────────────────────
...